locator_host/tools/mqtt_forwarder/requirements.txt
paho-mqtt==2.0.0
